tests of StatsToCSV StatsToExcel and documentation
