+++ requirements.txt
@@ -91,3 +91,4 @@ wcwidth==0.2.5; python_version >= "3.5"
 websocket-client==0.58.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
 zict==2.0.0; python_version >= "3.7"
 zipp==3.4.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.6.0" and python_version < "3.8" and python_version >= "3.6"
+openpyxl==3.0.7
